Fixed [Issue #2](https://github.com/yousefhany77/prisma-better-errors/issues/2) Added github actions to publish
