+++ .changeset/violet-rules-flow.md
@@ -1,0 +1,5 @@
+---
+"prisma-better-errors": patch
+---
+
+Fixed [Issue #2](https://github.com/yousefhany77/prisma-better-errors/issues/2) \n added github actions
