+++ src/index.ts
@@ -1,3 +1,1 @@
 export * from './prisma';
-export * from './prismaErros';
-export * from './util';
