+++ src/prisma/index.ts
@@ -1,3 +1,5 @@
+import { getPrismaErrorMessage } from './util';
+
 export declare class PrismaClientKnownRequestError extends Error implements ErrorWithBatchIndex {
   code: string;
   meta?: Record<string, unknown>;
@@ -17,3 +19,17 @@ declare type KnownErrorParams = {
   meta?: Record<string, unknown>;
   batchRequestIdx?: number;
 };
+
+export class prismaError extends Error {
+  statusCode: number;
+  title: string;
+  metaData?: string;
+
+  constructor(error: PrismaClientKnownRequestError) {
+    const { message, httpStatus } = getPrismaErrorMessage(error);
+    super(message);
+    this.title = 'Prisma Error';
+    this.statusCode = httpStatus;
+    this.metaData = error.meta && JSON.parse(JSON.stringify(error.meta));
+  }
+}
